refactor: update post list component
#70
