+++ src/app/post/page.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import ListFilter from '@/app/components/Post/ListFilter';
-import ListTable from '@/app/components/Post/ListTable';
+import List from '../components/Post/List';
 import Pagination from '@/app/components/Post/Pagination';
 import usePostList from '@/app/hooks/post/usePostList';
 import useReplaceSearchParams from '@/app/hooks/post/useReplaceSearchParams';
@@ -24,7 +24,7 @@ export default function PostList() {
     <section className="pb-8 pt-8">
       <ListFilter setSearchFilters={setSearchFilters} />
 
-      <ListTable posts={posts} isLoading={isFetching || isLoading || isError} />
+      <List posts={posts} isLoading={isFetching || isLoading || isError} />
 
       <Pagination meta={meta} setSearchFilters={setSearchFilters} />
     </section>
